agent_code/requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-huggingface==0.0.1
langchain-community==0.0.20
openai==1.12.0
python-dotenv==1.0.0

# Web Interface
gradio==4.16.0

# Web Scraping & Parsing
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
urllib3==2.1.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Excel Processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Additional Tools
youtube-transcript-api==0.6.2
arxiv==2.1.0
